テスト追加3

+++ timermute/db/mute_word_db.py
@@ -1,9 +1,10 @@
+from datetime import datetime
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.orm.exc import NoResultFound
 
 from timermute.db.base import Base
 from timermute.db.model import MuteWord
-from timermute.util import now
+from timermute.util import Result, now
 
 
 class MuteWordDB(Base):
@@ -17,14 +18,12 @@ class MuteWordDB(Base):
         session.close()
         return result
 
-    def upsert(self, record: str | MuteWord) -> int:
-        if isinstance(record, str):
-            keyword = str(record)
-            record = MuteWord(keyword, "muted", now(), now(), "")
+    def upsert(self, record: MuteWord) -> Result:
+        if not isinstance(record, MuteWord):
+            raise ValueError("record must be MuteWord.")
 
         Session = sessionmaker(bind=self.engine, autoflush=False)
         session = Session()
-        res = -1
 
         try:
             q = session.query(MuteWord).filter(MuteWord.keyword == record.keyword).with_for_update()
@@ -32,7 +31,6 @@ class MuteWordDB(Base):
         except NoResultFound:
             # INSERT
             session.add(record)
-            res = 0
         else:
             # UPDATE
             # id以外を更新する
@@ -41,45 +39,58 @@ class MuteWordDB(Base):
             p.created_at = record.created_at
             p.updated_at = record.updated_at
             p.unmuted_at = record.unmuted_at
-            res = 1
 
         session.commit()
         session.close()
-        return res
+        return Result.SUCCESS
 
-    def delete(self, key) -> None:
+    def delete(self, keyword: str) -> Result:
+        if not isinstance(keyword, str):
+            raise ValueError("keyword must be str.")
+
         Session = sessionmaker(bind=self.engine, autoflush=False)
         session = Session()
 
-        target = session.query(MuteWord).filter(MuteWord.keyword == key).one()
+        target = session.query(MuteWord).filter(MuteWord.keyword == keyword).one()
         session.delete(target)
 
         session.commit()
         session.close()
-        return
+        return Result.SUCCESS
 
-    def mute(self, key, unmuted_at) -> None:
+    def mute(self, keyword: str, unmuted_at: str) -> Result:
+        if not isinstance(keyword, str):
+            raise ValueError("keyword must be str.")
+        if not isinstance(unmuted_at, str):
+            raise ValueError("unmuted_at must be str.")
+
+        if unmuted_at:
+            destination_format = "%Y-%m-%d %H:%M:%S"
+            _ = datetime.strptime(unmuted_at, destination_format)
+
         Session = sessionmaker(bind=self.engine, autoflush=False)
         session = Session()
 
-        target = session.query(MuteWord).filter(MuteWord.keyword == key).one()
+        target = session.query(MuteWord).filter(MuteWord.keyword == keyword).one()
         target.status = "muted"
         target.updated_at = now()
         target.unmuted_at = unmuted_at
 
         session.commit()
         session.close()
-        return
+        return Result.SUCCESS
 
-    def unmute(self, key) -> None:
+    def unmute(self, keyword: str) -> Result:
+        if not isinstance(keyword, str):
+            raise ValueError("keyword must be str.")
         Session = sessionmaker(bind=self.engine, autoflush=False)
         session = Session()
 
-        target = session.query(MuteWord).filter(MuteWord.keyword == key).one()
+        target = session.query(MuteWord).filter(MuteWord.keyword == keyword).one()
         target.status = "unmuted"
         target.updated_at = now()
         target.unmuted_at = ""
 
         session.commit()
         session.close()
-        return
+        return Result.SUCCESS
